requirements.txt
python-telegram-bot==20.3
googletrans==4.0.0-rc1
pydub
gtts
SpeechRecognition